moves parquet and csv format configuration from data writers to destination
